Refatoração da repository, base inicial completa

+++ src/main/java/com/br/estimativadeprojetodesoftware/presenter/projeto/EdicaoProjetoPresenter.java
@@ -131,7 +131,7 @@ public class EdicaoProjetoPresenter {
             Map<String, Integer> tamanhosApp = perfil.getTamanhosApp();
 
             for (Map.Entry<String, Integer> entry : tamanhosApp.entrySet()) {
-                campos.add(new Campo(UUID.randomUUID(), "", entry.getKey(), entry.getValue()));
+                campos.add(new Campo(UUID.randomUUID(), "", entry.getKey(), entry.getValue().doubleValue()));
             }
 
             // Criando campos para niveisUD
@@ -143,7 +143,7 @@ public class EdicaoProjetoPresenter {
             // Criando campos para funcionalidades
             Map<String, Integer> funcionalidades = perfil.getFuncionalidades();
             for (Map.Entry<String, Integer> entry : funcionalidades.entrySet()) {
-                campos.add(new Campo(UUID.randomUUID(), "", entry.getKey(), entry.getValue()));
+                campos.add(new Campo(UUID.randomUUID(), "", entry.getKey(), entry.getValue().doubleValue()));
             }
 
             // Criando campos para taxas diarias
